requirements.txt
absl-py==1.4.0
accelerate==0.31.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-record==0.5.1
arrow==1.3.0
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.43.0
bleach==6.1.0
blinker==1.7.0
blobfile==2.1.0
cachetools==5.3.1
cfgv==3.3.1
chardet==5.2.0
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.6
clip==1.0
clu==0.0.12
cmake==3.29.0.1
colorama==0.4.6
comm==0.2.2
configargparse==1.7
contextlib2==21.6.0
contourpy==1.1.0
cycler==0.11.0
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==4.4.2
defusedxml==0.7.1
deprecated==1.2.14
diffusers==0.19.3
distlib==0.3.7
dm-tree==0.1.8
docker-pycreds==0.4.0
docstring-parser==0.15
easydict==1.13
einops==0.6.1
ema-pytorch==0.2.3
entrypoints==0.4
etils==1.7.0
exceptiongroup==1.2.0
executing==1.2.0
fastjsonschema==2.19.1
filelock==3.12.2
flask==3.0.2
flatbuffers==24.3.25
flax==0.8.2
fonttools==4.42.0
fqdn==1.5.1
frozendict==2.3.8
fsspec==2023.6.0
ftfy==6.1.1
gast==0.5.4
gdown==4.7.1
gitdb==4.0.10
gitpython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
hydra-core==1.3.2
icecream==2.1.3
identify==2.5.26
imageio==2.31.1
imageio-ffmpeg==0.4.8
immutabledict==4.2.0
importlib-metadata==6.8.0
importlib-resources==6.0.1
inplace-abn==1.1.0
ipycanvas==0.13.1
ipyevents==2.0.2
ipykernel==6.29.4
ipython==8.14.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jaxlib==0.4.27+cuda12.cudnn89
jaxtyping==0.2.28
jedi==0.19.0
joblib==1.4.0
json5==0.9.24
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==7.4.9
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter-server==2.13.0
jupyter-server-terminals==0.5.3
jupyterlab==4.1.5
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.4
jupyterlab-widgets==3.0.10
kaolin==0.15.0
keras==3.3.3
kiwisolver==1.4.4
kornia==0.6.0
lazy-loader==0.3
libclang==18.1.1
lightning-utilities==0.11.2
line-profiler==4.0.3
lit==18.1.2
lpips==0.1.4
lxml==4.9.4
markdown==3.4.4
markdown-it-py==3.0.0
marshmallow==3.21.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapy==1.1.9
mistune==3.0.2
ml-collections==0.1.1
ml-dtypes==0.4.0
moviepy==1.0.3
msgpack==1.0.5
munch==4.0.0
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
ninja==1.11.1.1
nodeenv==1.8.0
notebook-shim==0.2.4
numpy==1.26.4
nvdiffrast==0.3.1
nvidia-ml-py==12.535.77
nvitop==1.2.0
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.8.0.76
opt-einsum==3.3.0
optax==0.2.3.dev0
optree==0.11.0
orbax==0.1.9
orbax-checkpoint==0.5.9
overrides==7.7.0
packaging==23.1
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.4.0
platformdirs==4.2.1
plotly==5.20.0
plyfile==1.0.1
pre-commit==3.3.2
proglog==0.1.10
prometheus-client==0.20.0
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.12.0
pycryptodomex==3.20.0
pygltflib==1.16.2
pygments==2.16.1
pyliblzfse==0.4.1
pymcubes==0.1.4
pyopengl==3.1.7
pyparsing==3.0.9
pyquaternion==0.9.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
pywavelets==1.4.1
pyzmq==24.0.1
referencing==0.34.0
regex==2023.8.8
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.18.0
rsa==4.9
safetensors==0.3.2
scikit-image==0.21.0
scikit-learn==1.4.2
scipy==1.13.0
send2trash==1.8.2
sentencepiece==0.2.0
sentry-sdk==1.29.2
setproctitle==1.3.2
shapely==2.0.4
shtab==1.6.4
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.4.1
stack-data==0.6.2
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.1
tensorflow==2.16.1
tensorflow-datasets==4.9.4
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-metadata==1.15.0
tensorstore==0.1.56
termcolor==2.3.0
terminado==0.18.1
threadpoolctl==3.4.0
tifffile==2023.7.18
timm==0.9.5
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch-scatter==2.1.2
torchmetrics==1.1.1
torchsparse==1.4.0
torchtyping==0.1.4
tornado==6.4
tqdm==4.66.1
traitlets==5.9.0
transformers==4.31.0
trimesh==3.23.1
typeguard==2.13.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
tyro==0.5.5
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.18
usd-core==23.5
v==1
vector==1.3.1
vedo==2024.5.1
virtualenv==20.24.2
viser==0.0.16
vtk==9.3.0
wandb==0.15.8
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
werkzeug==3.0.2
widgetsnbextension==4.0.10
wrapt==1.16.0
zipp==3.16.2

